requirements.txt
Telethon==1.23.0
playsound==1.3.0
gTTS==2.2.3
python-dotenv==0.19.2